Implementando novas funções

+++ src/conta/Model/Bebida.java
@@ -1,7 +1,7 @@
 package conta.Model;
 
 public class Bebida extends ItemCardapio {
-    public Bebida(String nome, double preco) {
+    public Bebida(String nome, float preco) {
         super(nome, preco);
         
     }
