+++ src/conta/Model/Marmita.java
@@ -3,7 +3,7 @@ package conta.Model;
 //import conta.ItemCardapio;
 
 public class Marmita extends ItemCardapio {
-    public Marmita(String nome, double preco) {
+    public Marmita(String nome, float preco) {
         super(nome, preco);
     }
 }
